Add files via upload
Instruction on pdf. a lot of the errors were bc of not matching the spaces and specific definitions correctly, look at the behavior of what the output should do and make sure that the methods inside the class matches that. Also note the use of the super.init__ bc you need this to override the parent class (animal) class behavior for the dog method